+++ src/locale/en/notification/index.ts
@@ -4,9 +4,11 @@ export const notification = {
       first: 'Main',
       second: 'Description',
       third: 'Users',
+      forth: 'Other',
       mainInfo: 'Main info',
       description: 'Description',
       users: 'Users and groups',
+      groupsRoles: 'Groups and roles',
     },
     title: 'Title',
     author: 'Author',
